Added Awnsering Questions
In the main routiene I have a section thay displays all the 10 catagories they can choose from. The player now enters what catagory they want to play for the round and how many questions pp. The play turn function in the Player class is then called. This function displays the question and awnsers below in a random order and the user can awnser the question. The program then displays if they are correct or not. I have also chaned the get_int function and removed the paramater. Now when the function is called > will be displayed. I did this as I dont want to display and other masage.