Added the scoring system
I have added a scoring system that gives a player a point for every correct awnser. At the end of the game, each players score and correct pecentage are displayed using the print_players_stats function. From this a winner is found with the max score and diaplayed. I have also added more input validation to the API call so it now doesn't crash when you lose wifi, fixed it so the correct option corrosponds to the correct awnser, printed out the correct awnser when you get it wrong, and allowed to user to play any number of questions 1 - 5.