Bug fixes and extra features
When " where a character in the question they came out as a string of letters. Used html import  and .unescape function to fix this. Made is so each player gets the same ammount of a difficulty of question. Added name = main to define main routiene. Fixed a bug in call_api() function so now when too many questions have been asked it works.